User- Subscription relationship

+++ app/Http/Controllers/RegisterController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Invite;
+use App\Subscriber;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
@@ -19,6 +20,11 @@ class RegisterController extends Controller
         return view('register');
     }
 
+    protected function checkEmail()
+    {
+
+    }
+
     protected function create(Request $request)
     {
         $this->validate($request, [
@@ -28,11 +34,16 @@ class RegisterController extends Controller
             'token'    => 'required|string|exists:invites,token',
         ]);
 
-        User::create([
-            'name'     => request('name'),
-            'email'    => request('email'),
-            'password' => Hash::make(request('password')),
-            'type'     => User::DEFAULT_TYPE,
+        if ( ! $subscriber = Subscriber::where('email', $request->email)->first()) {
+            return back()->with('status', 'Please sign in with the email that you gave to us in order to invite you');
+        }
+
+        $user = User::create([
+            'name'            => request('name'),
+            'email'           => request('email'),
+            'subscription_id' => $subscriber->id,
+            'password'        => Hash::make(request('password')),
+            'type'            => User::DEFAULT_TYPE,
         ]);
 
         Invite::where('token', '=', request('token'))->delete(); //deletes invitation(with token) after registration
